+++ README.md
@@ -109,8 +109,18 @@
 ##### ➂このアプリの（ユーザ）サイトマップ
 ![このアプリの（ユーザ）サイトマップ](/public/uploads/sample5.png)
 
+## 4. このWebアプリケーションを開発した経緯
+テックアカデミーでのレッスンの集大成として、このオリジナルアプリを開発しました。
+友人がダンススクールの運営をしていて、「お客様の予約の管理をLineで一人一人対応している為大変！」というお話を聞いたのがきっかけで、レッスンの予約をアプリでできたら便利だろうなと思い開発しました。
+ダンススクールの予約サイトとして、業務の効率化を高める目的として、このアプリ開発に取り組みました。
+
+## 5. お問い合わせ
+駆け出しエンジニアの立場で、まだまだ不勉強なためバグが潜んでいるかもしれません。
+現在、ユーザの退会、キャンセル機能は前日までに設定するなどの作成に挑戦しおります。
+改善点などがありましたら、以下のメールアドレスにご連絡いただけると幸いです。
+
 ##### ◆メールアドレス:
 misato71h@gmail.com
 
 ## 著者
-2022/08/18 Misato Harada
+2022/08/18 Harada Misato
