+++ app/Http/Controllers/InstructorsController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
-use App\Models\Instructor;
+use App\Instructor;
 
 class InstructorsController extends Controller
 {
