+++ app/Http/Controllers/LessonsController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
-use App\Models\Lesson;
+use App\Lesson;
 
 class LessonsController extends Controller
 {
