+++ app/Http/Controllers/StudiosController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
-use App\Models\Studio;
+use App\Studio;
 
 class StudiosController extends Controller
 {
