+++ app/Http/Controllers/UsersController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Models\User;
+use App\User;
 
 class UsersController extends Controller
 {
