+++ composer.json
@@ -37,8 +37,7 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/",
-            "Models\\": "app/Models/"
+            "App\\": "app/"
         },
         "classmap": [
             "database/seeds",
